+++ src/Sql.Cmdlets/Invoke-Query.cs
@@ -7,14 +7,14 @@ using System.Reflection;
 /// <summary>
 /// Executes a parameterized SQL query and returns an array of objects whose properties correspond to the columns.
 /// </summary>
-[Cmdlet(VerbsLifecycle.Invoke, "Query"), OutputType(typeof(object))]
+[Cmdlet(VerbsLifecycle.Invoke, "Query"), OutputType(typeof(object), typeof(Tuple<object?, object?>))]
 public class InvokeQueryCommand: Cmdlet {
 
 	/// <summary>
 	/// The type of objects to return.
 	/// </summary>
-	[Parameter]
-	public Type As { get; set; } = typeof(ExpandoObject);
+	[Parameter, ValidateNotNullOrEmpty]
+	public Type[] As { get; set; } = [typeof(ExpandoObject)];
 
 	/// <summary>
 	/// The SQL query to be executed.
@@ -41,6 +41,12 @@ public class InvokeQueryCommand: Cmdlet {
 	public ParameterCollection Parameters { get; set; } = [];
 
 	/// <summary>
+	/// The field from which to split and read a second object.
+	/// </summary>
+	[Parameter, ValidateNotNullOrWhiteSpace]
+	public string SplitOn { get; set; } = "Id";
+
+	/// <summary>
 	/// The wait time, in seconds, before terminating the attempt to execute the command and generating an error.
 	/// </summary>
 	[Parameter, ValidateRange(ValidateRangeKind.Positive)]
@@ -56,11 +62,17 @@ public class InvokeQueryCommand: Cmdlet {
 	/// Performs execution of this command.
 	/// </summary>
 	protected override void ProcessRecord() {
+		Type[] types = As.Length <= 1
+			? [typeof(IDbConnection), typeof(string), typeof(ParameterCollection), typeof(CommandOptions)]
+			: [typeof(IDbConnection), typeof(string), typeof(ParameterCollection), typeof(string), typeof(CommandOptions)];
+
+		object?[] arguments = As.Length <= 1
+			? [Connection, Command, Parameters, new CommandOptions(Timeout, Transaction, CommandType)]
+			: [Connection, Command, Parameters, SplitOn, new CommandOptions(Timeout, Transaction, CommandType)];
+
 		try {
-			var types = new[] { typeof(IDbConnection), typeof(string), typeof(ParameterCollection), typeof(CommandOptions) };
-			var method = typeof(ConnectionExtensions).GetMethod(nameof(ConnectionExtensions.Query), 1, types)!.MakeGenericMethod(As);
-			var records = (IEnumerable<object>) method.Invoke(null, [Connection, Command, Parameters, new CommandOptions(Timeout, Transaction, CommandType)])!;
-			WriteObject(records, enumerateCollection: true);
+			var method = typeof(ConnectionExtensions).GetMethod(nameof(ConnectionExtensions.Query), As.Length, types)!.MakeGenericMethod(As);
+			WriteObject(method.Invoke(null, arguments), enumerateCollection: true);
 		}
 		catch (TargetInvocationException e) {
 			WriteError(new ErrorRecord(e.InnerException, "Invoke-Query:TargetInvocationException", ErrorCategory.OperationStopped, null));